Emit correct 'flags' type from FilesModel::clickedItem()
Saves a lot of verbose static_cast()'ing to and from
QItemSelectionModel::SelectionFlags.

+++ src/core/filesmodel.cpp
@@ -272,7 +272,7 @@ void FilesModel::clickedPoint(const TrackDataAbstractPoint *tdp, Qt::KeyboardMod
     else return;
 
     qDebug() << "click for" << indexForItem(tdp) << "flags" << selFlags;
-    emit clickedItem(indexForItem(tdp), static_cast<unsigned int>(selFlags));
+    emit clickedItem(indexForItem(tdp), selFlags);
 }
 
 //////////////////////////////////////////////////////////////////////////
